Fix test_ddr5_cs_n_multiple_phases
Changed sampling clock for `cs_n` to `sys4x`. It shouldn't be shifted
in phase.

Removed part of `dfi_sequence` that checked behaviour of overlapping
commands, as the DFI now assumes it gets commands in correct phases and
commands do not overlap.

Instead added `test_ddr5_cs_n_overlapping_phases`.
This test checks that when we issue a 2 cycle command on phase 2 and
another command on phase 3, then it will override the previous one.
`cs_n` value will be set to 0, but `ca` pads` are not checked as it is
an undefined behaviour.

Also added vcd trace generation for debugging purposes.

Signed-off-by: Michal Sieron <msieron@antmicro.com>