Fix test_ddr5_ca_addressing
Changed sampling clock for `cs_n` and `ca` to `sys4x`. It shouldn't
be shifted in phase.

Due to changes in the DFI injectors, commands are now passed via
`address` key in the DFI command dictionary.
Also command translation no longer handles two-cycle commands for us.
That's why we need to provide each part of the command separately in
appropriate DFI phase.
That's where changes in commands come from.

Renamed `write_ap` to `write` as `phy/ddr5/commands.py` sets CA10 high,
which means it's a normal 'Write' command and not 'Write with Auto
Precharge'.

In `pad_checkers` I had to modify expected value on `ca5` in read and
write commands, as it is related to burst length (we only support BL8)
so it has to be 0.

Also added vcd trace generation for debugging purposes.

Signed-off-by: Michal Sieron <msieron@antmicro.com>